app crashing while adding amount in text field resolved,
changed amount datatype from double to BigDecimal

and some ui issue resolved

+++ app/src/main/java/com/ekspensify/app/data/model/responseModel/AccountSummaryModel.kt
@@ -3,16 +3,17 @@ package com.ekspensify.app.data.model.responseModel
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
 import java.io.Serializable
+import java.math.BigDecimal
 
 data class AcSummaryResponseModel(
     @SerializedName("total")
     @Expose
-    var total: Double? = null,
+    var total: BigDecimal? = null,
     @SerializedName("credit")
     @Expose
-    var credit: Double? = null,
+    var credit: BigDecimal? = null,
 
     @SerializedName("debit")
     @Expose
-    var debit: Double? = null
+    var debit: BigDecimal? = null
 ) : Serializable
